Move User in tests to the `tobuilder` package.
